+++ lib/domain/repositories/auth_repository.dart
@@ -4,8 +4,8 @@ import '../../core/errors/failures.dart';
 import '../entities/user.dart';
 
 abstract class AuthRepository {
-  Future<Either<Failure, User>> signInWithSchoolCode({
-    required String schoolCode,
+  /// Sign in with student number (globally unique) and password
+  Future<Either<Failure, User>> signInWithStudentNumber({
     required String studentNumber,
     required String password,
   });
@@ -18,8 +18,6 @@ abstract class AuthRepository {
   Future<Either<Failure, void>> signOut();
 
   Future<Either<Failure, User?>> getCurrentUser();
-
-  Future<Either<Failure, bool>> validateSchoolCode(String code);
 
   Stream<User?> get authStateChanges;